% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preText_test.R
\name{get_max_difference}
\alias{get_max_difference}
\title{Get ordering of pairwise distances by maximum difference from baseline}
\usage{
get_max_difference(dist_matrix, baseline)
}
\arguments{
\item{dist_matrix}{A distance matrix.}

\item{baseline}{A baseline distance matrix.}
}
\value{
An integer vector of indices ordered by absolute difference.
}
\description{
Get ordering of pairwise distances by maximum difference from baseline
}
\keyword{internal}
