% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{preText2-package}
\alias{preText2-package}
\alias{preText2}
\title{preText2: Diagnostics to Assess the Effects of Text Preprocessing Decisions}
\description{
An updated fork of the preText package, compatible with quanteda v4+.
Functions to assess the effects of different text preprocessing decisions
on the inferences drawn from the resulting document-term matrices.
}
\details{
The main functions are:
\itemize{
  \item \code{\link{factorial_preprocessing}}: Preprocess text 64 or 128 ways
  \item \code{\link{preText}}: Run the full preText procedure
  \item \code{\link{preText_score_plot}}: Plot preText scores
  \item \code{\link{regression_coefficient_plot}}: Plot regression coefficients
}
}
\references{
Matthew J. Denny, and Arthur Spirling (2018). "Text Preprocessing For
Unsupervised Learning: Why It Matters, When It Misleads, And What To Do
About It". Political Analysis, 26(2), 168-189.
\doi{10.1017/pan.2017.44}
}
\author{
\strong{Maintainer}: Arthur Spirling \email{as9934@nyu.edu}

Authors:
\itemize{
  \item Matthew J. Denny \email{mdenny@psu.edu}
}

}
