% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{preText_score_plot}
\alias{preText_score_plot}
\title{preText specification plot}
\usage{
preText_score_plot(preText_results, remove_intercept = TRUE)
}
\arguments{
\item{preText_results}{A results object from the \code{preText()} function.}

\item{remove_intercept}{Logical, not used for this plot but included for
API compatibility.}
}
\value{
A ggplot object.
}
\description{
Creates a dot plot of preText scores for each preprocessing
specification, ordered from most to least unusual.
}
\examples{
\dontrun{
preText_score_plot(preText_results)
}
}
