% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preText_test.R
\name{preText_test}
\alias{preText_test}
\title{preText Test}
\usage{
preText_test(
  distance_matrices,
  choices,
  labels = NULL,
  baseline_index = 128,
  text_size = 1,
  num_comparisons = 50,
  parallel = FALSE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{distance_matrices}{A list of document distance matrices generated by the
\code{scaling_comparison()} function.}

\item{choices}{A dataframe indicating whether a preprocessing step was
applied or not. Generated by \code{factorial_preprocessing()}.}

\item{labels}{Optional labels for each preprocessing step.}

\item{baseline_index}{The index of the baseline distance matrix.
Defaults to 128.}

\item{text_size}{The cex for text in plots. Defaults to 1.}

\item{num_comparisons}{The number of ranks to use. Defaults to 50.}

\item{parallel}{Logical for parallel processing. Defaults to FALSE.}

\item{cores}{Number of cores. Defaults to 1.}

\item{verbose}{Logical for printing progress. Defaults to TRUE.}
}
\value{
A result list with \code{preText_scores} and
\code{ranked_preText_scores}.
}
\description{
Calculates preText scores for each preprocessing specification
by comparing document distance matrices.
}
\examples{
\dontrun{
# This function is typically called inside preText().
}
}
