% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{regression_coefficient_plot}
\alias{regression_coefficient_plot}
\title{Regression Coefficient Plot}
\usage{
regression_coefficient_plot(preText_results, remove_intercept = TRUE)
}
\arguments{
\item{preText_results}{A results object from the \code{preText()} function.}

\item{remove_intercept}{Logical indicating whether to remove the intercept
from the plot. Defaults to TRUE.}
}
\value{
A ggplot object.
}
\description{
Creates a dot plot of regression coefficients showing the
effect of each preprocessing decision on preText scores.
}
\examples{
\dontrun{
regression_coefficient_plot(preText_results, remove_intercept = TRUE)
}
}
