% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_infrequent_terms.R
\name{remove_infrequent_terms}
\alias{remove_infrequent_terms}
\title{Remove infrequently occurring terms from quanteda dfm.}
\usage{
remove_infrequent_terms(
  dfm_object,
  proportion_threshold = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{dfm_object}{A quanteda dfm object.}

\item{proportion_threshold}{The proportion of documents a term must appear
in to be retained. Defaults to 0.01.}

\item{verbose}{Logical indicating whether to print information. Defaults to
TRUE.}
}
\value{
A reduced dfm.
}
\description{
Removes terms appearing in less than a specific proportion of
documents in a corpus from a dfm.
}
\examples{
\dontrun{
library(preText2)
data("UK_Manifestos")
corp <- quanteda::corpus(UK_Manifestos)
toks <- quanteda::tokens(corp, remove_punct = TRUE)
my_dfm <- quanteda::dfm(toks)
reduced_dfm <- remove_infrequent_terms(my_dfm, 0.05)
}
}
