% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{data}
\name{UK_Manifestos}
\alias{UK_Manifestos}
\title{Full text of 69 UK party manifestos from 1918-2001.}
\format{
A character vector of length 69.
}
\usage{
data(UK_Manifestos)
}
\description{
A character vector containing 69 entries, one per document, for all
major party manifestos in the UK from 1918 to 2001.
}
\keyword{datasets}
