% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{preText_score_plot}
\alias{preText_score_plot}
\title{preText specification plot}
\usage{
preText_score_plot(preText_results, remove_constant = TRUE)
}
\arguments{
\item{preText_results}{A list object returned by the \code{preText()} function.}

\item{remove_constant}{Logical; remove specifications with zero score.
Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Generates a dot plot of preText scores for each preprocessing
  specification, ordered from most to least unusual.
}
