% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{regression_coefficient_plot}
\alias{regression_coefficient_plot}
\title{Regression Coefficient Plot}
\usage{
regression_coefficient_plot(preText_results, remove_intercept = TRUE)
}
\arguments{
\item{preText_results}{A list object returned by the \code{preText()} function.}

\item{remove_intercept}{Logical; whether to remove the intercept from the
plot. Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Generates a coefficient plot showing the effect of each
  preprocessing step on the preText score.
}
